backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.12.0
shap==0.44.1
python-multipart==0.0.6
